requirements.txt
PyGithub==1.43.7
pickleDB==0.9.2
requests==2.22.0
